+++ example/angular2-try/components/index.js
@@ -1,5 +1,5 @@
-System.register(['angular2/router', 'angular2/angular2', './app', '../services/common.service'], function(exports_1) {
-    var router_1, angular2_1, app_1, common_service_1;
+System.register(['angular2/router', 'angular2/angular2', 'angular2/di', './app', '../services/common.service'], function(exports_1) {
+    var router_1, angular2_1, di_1, app_1, common_service_1;
     return {
         setters:[
             function (_router_1) {
@@ -8,6 +8,9 @@ System.register(['angular2/router', 'angular2/angular2', './app', '../services/c
             function (_angular2_1) {
                 angular2_1 = _angular2_1;
             },
+            function (_di_1) {
+                di_1 = _di_1;
+            },
             function (_app_1) {
                 app_1 = _app_1;
             },
@@ -15,7 +18,11 @@ System.register(['angular2/router', 'angular2/angular2', './app', '../services/c
                 common_service_1 = _common_service_1;
             }],
         execute: function() {
-            angular2_1.bootstrap(app_1.App, [router_1.routerInjectables, common_service_1.CommonService, angular2_1.bind(router_1.LocationStrategy).toClass(router_1.HTML5LocationStrategy)]);
+            angular2_1.bootstrap(app_1.App, [
+                router_1.routerInjectables,
+                common_service_1.CommonService,
+                di_1.bind(router_1.LocationStrategy).toClass(router_1.HTML5LocationStrategy)
+            ]);
         }
     }
 });