+++ example/angular2-try/components/index.ts
@@ -1,7 +1,12 @@
 import {routerInjectables, LocationStrategy, HashLocationStrategy, HTML5LocationStrategy} from 'angular2/router';
-import {bootstrap, bind} from 'angular2/angular2';
+import {bootstrap} from 'angular2/angular2';
+import {bind} from 'angular2/di';
+
 import {App} from './app';
 import {CommonService} from '../services/common.service';
 
-
-bootstrap(App, [routerInjectables, CommonService, bind(LocationStrategy).toClass(HTML5LocationStrategy)]);
+bootstrap(App, [
+    routerInjectables,
+    CommonService,
+    bind(LocationStrategy).toClass(HTML5LocationStrategy)
+]);