+++ example/angular2-try/directives/greeter.js
@@ -10,7 +10,7 @@ var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };System.register(['angular2/angular2', 'angular2/di'], function(exports_1) {
     var angular2_1, di_1;
-    var Highlight, RedDec, Tooltip, NeedsGreeter, Greeter;
+    var RedDec, Greeter;
     return {
         setters:[
             function (_angular2_1) {
@@ -20,78 +20,62 @@ var __metadata = (this && this.__metadata) || function (k, v) {
                 di_1 = _di_1;
             }],
         execute: function() {
-            Highlight = (function () {
-                function Highlight(element) {
-                    this.element = element;
-                }
-                Highlight.prototype.onMouseEnter = function () {
-                    this.outline('#f00 solid 1px');
-                };
-                Highlight.prototype.onMouseLeave = function () {
-                    this.outline();
-                };
-                Highlight.prototype.outline = function (outline) {
-                    if (outline === void 0) { outline = ""; }
-                    this.element.domElement.style.outline = outline;
-                };
-                Highlight = __decorate([
-                    angular2_1.Directive({
-                        selector: 'hightlight',
-                        hostListeners: {
-                            'mouseenter': 'onMouseEnter()',
-                            'mouseleave': 'onMouseLeave()'
-                        }
-                    }), 
-                    __metadata('design:paramtypes', [(typeof ElementRef !== 'undefined' && ElementRef) || Object])
-                ], Highlight);
-                return Highlight;
-            })();
-            exports_1("Highlight", Highlight);
+            //@Directive({
+            //    selector: 'hightlight',
+            //    hostListeners: {
+            //        'mouseenter': 'onMouseEnter()',
+            //        'mouseleave': 'onMouseLeave()'
+            //    }
+            //})
+            //export class Highlight {
+            //
+            //    constructor(private element:ElementRef) {
+            //    }
+            //
+            //    onMouseEnter() {
+            //        this.outline('#f00 solid 1px');
+            //    }
+            //
+            //    onMouseLeave() {
+            //        this.outline();
+            //    }
+            //
+            //    private outline(outline:string = "") {
+            //        this.element.domElement.style.outline = outline;
+            //    }
+            //}
             RedDec = (function () {
-                function RedDec(el) {
-                    el.domElement.style.color = 'red';
+                // ElementRef is always injectable and it wraps the element on which the
+                // directive was found by the compiler.
+                function RedDec(el, renderer) {
+                    renderer.setElementStyle(el, 'color', 'red');
                 }
                 RedDec = __decorate([
-                    angular2_1.Directive({ selector: 'red' }), 
-                    __metadata('design:paramtypes', [(typeof ElementRef !== 'undefined' && ElementRef) || Object])
+                    angular2_1.Directive({ selector: '[red]' }), 
+                    __metadata('design:paramtypes', [(typeof ElementRef !== 'undefined' && ElementRef) || Object, (typeof Renderer !== 'undefined' && Renderer) || Object])
                 ], RedDec);
                 return RedDec;
             })();
-            exports_1("RedDec", RedDec);
-            Tooltip = (function () {
-                function Tooltip() {
-                }
-                Object.defineProperty(Tooltip.prototype, "text", {
-                    set: function (value) {
-                        console.log("Tooltip Directive new value: " + value);
-                    },
-                    enumerable: true,
-                    configurable: true
-                });
-                Tooltip = __decorate([
-                    angular2_1.Directive({
-                        selector: '[tooltip]',
-                        properties: ['text: tooltip']
-                    }), 
-                    __metadata('design:paramtypes', [])
-                ], Tooltip);
-                return Tooltip;
-            })();
-            exports_1("Tooltip", Tooltip);
-            NeedsGreeter = (function () {
-                //constructor(greeter: Greeter) {
-                //    this.greeter = greeter;
-                //}
-                function NeedsGreeter() {
-                    //this.greeter = greeter;
-                }
-                NeedsGreeter = __decorate([
-                    angular2_1.Directive({ selector: 'needs-greeter' }), 
-                    __metadata('design:paramtypes', [])
-                ], NeedsGreeter);
-                return NeedsGreeter;
-            })();
-            exports_1("NeedsGreeter", NeedsGreeter);
+            //@Directive({
+            //    selector: '[tooltip]',
+            //    properties: ['text: tooltip']
+            //})
+            //export class Tooltip {
+            //    set text(value: string) {
+            //        console.log("Tooltip Directive new value: " + value);
+            //    }
+            //}
+            //@Directive({selector: 'needs-greeter'})
+            //export class NeedsGreeter {
+            //  greeter: Greeter;
+            //
+            //  //constructor(greeter: Greeter) {
+            //  //    this.greeter = greeter;
+            //  //}
+            //    constructor() {
+            //        //this.greeter = greeter;
+            //    }
+            //}
             Greeter = (function () {
                 function Greeter() {
                 }